Unicode fix and syntax line
